Move epoch processing benches into separate file
